requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
decorator==4.4.2
exceptiongroup==1.1.1
fastapi==0.97.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.12.2
frozenlist==1.3.3
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-texttospeech==2.14.1
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
httplib2==0.22.0
idna==3.4
imageio==2.31.1
imageio-ffmpeg==0.4.8
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.2
moviepy==1.0.3
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
Pillow==9.5.0
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.23.2
py-dotenv==0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
tinysegmenter==0.3
tldextract==3.4.4
tqdm==4.65.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
yarl==1.9.2
